refactor: rename components to better reflect their content
